+++ civiclens/nlp/comments.py
@@ -3,7 +3,7 @@ import polars as pl
 from networkx.algorithms.community import louvain_communities
 from sentence_transformers import SentenceTransformer, util
 
-from ..utils.database_access import pull_data
+from civiclens.utils.database_access import pull_data
 
 
 def get_doc_comments(schema: list[str]) -> pl.DataFrame:
