+++ civiclens/nlp/tools.py
@@ -1,7 +1,62 @@
+from datetime import datetime
+
 import polars as pl
 import torch
+from pydantic import BaseModel
 from sentence_transformers.util import cos_sim
 from transformers import pipeline
+
+
+class Comments(BaseModel, arbitrary_types_allowed=True):
+    df: pl.DataFrame
+    document: str = ""
+    rep_comments: list[str] = []
+    num_comments: int = 0
+    num_unique_comments: int = 0
+    num_rep_comments: int = 0
+    topics: dict = {}
+    created_at: datetime = datetime.today()
+    _comment_list: list = []
+
+    def __init__(self, *args, **kwargs):
+        # assigns other values
+        super().__init__(*args, **kwargs)
+        self.num_comments = len(self.df)
+
+    def to_list(self, field: str = "Comment"):
+        """
+        Returns list of comments
+        """
+        if not self._comment_list:
+            self._comment_list = self.df[field].to_list()
+
+        return self._comment_list
+
+    def __len__(self):
+        """
+        Return number of total comments
+        """
+        return self.num_comments
+
+
+class Comment(BaseModel):
+    # probably should use id from API or
+    uuid: str = ""
+    text: str = ""
+    comments_represented: int = 0
+    representative: bool = False
+    form_letter: bool = False
+
+
+def comments_from_sql():
+    pass
+
+
+def comments_from_df(df: pl.DataFrame) -> list[Comment]:
+    """
+    Creates list of comments of from
+    """
+    pass
 
 
 def extract_formletters(
